requirements.txt
request
mongoengine
pymongo
SolrClient
nltk
pandas
googletrans
pysolr